requirements.txt
Scrapy
django
celery[redis]
uvicorn
mysqlclient
sqlalchemy
djangorestframework